fix some warning

+++ core/include/resample.h
@@ -13,10 +13,10 @@ struct SwrContextParam
 {
 	int64_t				out_ch_layout;
 	enum AVSampleFormat out_sample_fmt;
-	int					out_sample_rate;
+	int64_t				out_sample_rate;
 	int64_t				in_ch_layout;
 	enum AVSampleFormat in_sample_fmt;
-	int					in_sample_rate;
+	int64_t				in_sample_rate;
 	int					log_offset;
 	void*				log_ctx;
 	int					fullOutputBufferSize = -1;
